@@ -1,11 +1,12 @@
 {
   "name": "@xialvjun/mapbox-gl-vue",
-  "version": "0.1.2",
+  "version": "0.2.0",
   "description": "a simple mapbox-gl vue bind, compatible for both 2 and 3",
   "main": "index.js",
   "scripts": {
     "build": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test2": "vue-demi-switch 2 vue2 && parcel --no-cache demos/vue2/index.html",
+    "test3": "vue-demi-switch 3 && parcel --no-cache demos/vue3/index.html"
   },
   "repository": {
     "type": "git",
@@ -21,12 +22,26 @@
     "url": "https://github.com/xialvjun/mapbox-gl-vue/issues"
   },
   "homepage": "https://github.com/xialvjun/mapbox-gl-vue#readme",
+  "dependencies": {
+    "vue-demi": "latest"
+  },
   "peerDependencies": {
+    "@vue/composition-api": "^1.0.0",
     "mapbox-gl": "^2.0.0",
-    "vue": ">=2.5.6"
+    "vue": "^2.5.0 || >=3.0.0"
+  },
+  "peerDependenciesMeta": {
+    "@vue/composition-api": {
+      "optional": true
+    }
   },
   "devDependencies": {
     "@types/mapbox-gl": "^2.0.0",
-    "typescript": "^4.1.3"
+    "@vue/composition-api": "^1.6.2",
+    "mapbox-gl": "^2.8.2",
+    "parcel": "^2.6.0",
+    "typescript": "^4.1.3",
+    "vue": "^3.0.0",
+    "vue2": "npm:vue@2"
   }
 }
