@@ -1,9 +1,9 @@
-import Vue from "vue";
-import * as coms from "./coms";
+import Vue from 'vue2';
+import { MglMap, MglOriginMarker, MglImages, MglGeojsonSource, MglLayer } from '../../index';
 
 const map_init = {
   accessToken: `pk.eyJ1IjoiZG9sYnl6ZXJyIiwiYSI6InhIS25oN0EifQ.QQrwwFUZu6trJNjGsrpTFQ`,
-  style: "mapbox://styles/mapbox/streets-v11",
+  style: 'mapbox://styles/mapbox/streets-v11',
   center: [116.396119, 39.925617] as const,
   zoom: 10,
 };
@@ -11,15 +11,15 @@ const img = `https://avatars1.githubusercontent.com/u/4413712?s=60&v=4`;
 
 let App = Vue.extend({
   render(h) {
-    return h("div", { style: "width:500px;height:500px" }, [h(coms.Map, { props: { init: map_init } })]);
+    return h('div', { style: 'width:500px;height:500px' }, [h(MglMap, { props: { init: map_init } })]);
   },
 });
 
 App = Vue.extend({
   render(h) {
-    return h("div", { style: "width:500px;height:500px" }, [
-      h(coms.Map, { props: { init: map_init } }, [
-        h(coms.OriginMarker, { props: { lng_lat: [116.396119, 39.925617] } }, [h("img", { domProps: { src: img } })]),
+    return h('div', { style: 'width:500px;height:500px' }, [
+      h(MglMap, { props: { init: map_init } }, [
+        h(MglOriginMarker, { props: { lng_lat: [116.396119, 39.925617] } }, [h('img', { domProps: { src: img } })]),
       ]),
     ]);
   },
@@ -30,45 +30,45 @@ App = Vue.extend({
     return { marker: true, images: true, geojson: true };
   },
   render(h) {
-    return h("div", [
-      h("button", { on: { click: () => (this.marker = !this.marker) } }, "marker" + this.marker),
-      h("button", { on: { click: () => (this.images = !this.images) } }, "images"+this.images),
-      h("button", { on: { click: () => (this.geojson = !this.geojson) } }, "geojson"+this.geojson),
-      h("div", { style: "width:500px;height:500px" }, [
-        h(coms.Map, { props: { init: map_init } }, [
-          this.marker && h(coms.OriginMarker, { props: { lng_lat: [116.396119, 39.925617] } }, [h("img", { domProps: { src: img } })]),
+    return h('div', [
+      h('button', { on: { click: () => (this.marker = !this.marker) } }, 'marker' + this.marker),
+      h('button', { on: { click: () => (this.images = !this.images) } }, 'images' + this.images),
+      h('button', { on: { click: () => (this.geojson = !this.geojson) } }, 'geojson' + this.geojson),
+      h('div', { style: 'width:500px;height:500px' }, [
+        h(MglMap, { props: { init: map_init } }, [
+          this.marker && h(MglOriginMarker, { props: { lng_lat: [116.396119, 39.925617] } }, [h('img', { domProps: { src: img } })]),
           this.images &&
-            h(coms.Images, { props: { init_imgs: { xxx: img } } }, [
+            h(MglImages, { props: { init_imgs: { xxx: img } } }, [
               this.geojson &&
                 h(
-                  coms.GeojsonSource,
+                  MglGeojsonSource,
                   {
                     props: {
                       data: {
-                        type: "Feature",
+                        type: 'Feature',
                         properties: {
-                          icon: "theatre",
+                          icon: 'theatre',
                         },
                         geometry: {
-                          type: "Point",
+                          type: 'Point',
                           coordinates: [116.396119, 39.925617],
                         },
                       },
                     },
                   },
                   [
-                    h(coms.Layer, {
+                    h(MglLayer, {
                       props: {
                         layer: {
-                          type: "symbol",
+                          type: 'symbol',
                           layout: {
-                            "icon-image": "xxx",
-                            "icon-allow-overlap": true,
+                            'icon-image': 'xxx',
+                            'icon-allow-overlap': true,
                           },
                         },
                       },
                     }),
-                  ],
+                  ]
                 ),
             ]),
         ]),
@@ -77,4 +77,4 @@ App = Vue.extend({
   },
 });
 
-new App().$mount("#app");
+new App().$mount('#app');
